Added a missing header and compiler warnings.
Also let the auto-formatter cleanup a few of the files
that I was in.

+++ bbs/diredit.cpp
@@ -374,7 +374,7 @@ void delete_dir(int n) {
     m2 = 0xffffffff << (n % 32);
     m3 = 0xffffffff >> (32 - (n % 32));
 
-    for (size_t i = 1; i <= num_users; i++) {
+    for (int i = 1; i <= num_users; i++) {
       read_qscn(i, pTempQScan, true);
 
       pTempQScan_n[n / 32] = (pTempQScan_n[n / 32] & m3) | ((pTempQScan_n[n / 32] >> 1) & m2) |
